# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Phantom is a web application that allows passengers to track bus location and movements around Kigali",
    "version": "1.0",
    "title": "Phantom",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "email": "isaackomeza@gmail.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://github.com/atlp-rwanda/orcas-phantom-backend/README.md"
    }
  },
  "host": "phantom-backend.herokuapp.com",
  "basePath": "/",
  "schemes": ["http", "https"],
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "paths": {
...
    "/api/users/{id}": {
      "get": {
        "tags": ["users"],
        "summary": "Admin get a specific user",
        "description": "This endpoint uses get request to get a specific user by id",
        "operationId": "getSpecificUser",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "user object that need get user by id",
            "required": true
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "type": "string",
            "description": "Admin Token"
          }
        ],
        "responses": {
          "200": {
            "description": "user retrieved"
          },
          "401": {
            "description": "You are not authorized to use this resource"
          },
          "404": {
            "description": "no available user"
          },
          "500": {
            "description": "internal server error"
          }
        }
      },
      "patch": {
        "tags": ["users"],
        "summary": "Admin update a specific user",
        "description": "This endpoint uses patch request to update a specific user by id",
        "operationId": "updateUser",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "user object that need to update a user by id",
            "required": true
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "type": "string",
            "description": "Admin Token"
          },
          {
            "description": "user object login use",
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UserInput"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User was updated successfully."
          },
          "401": {
            "description": "You are not authorized to use this resource"
          },
          "404": {
            "description": "There is no available user!"
          },
          "409": {
            "description": "Email address already taken"
          },
          "500": {
            "description": "internal server error"
          }
        }
      },
      "delete": {
        "tags": ["users"],
        "summary": "Admin delete a specific user",
        "description": "This endpoint uses delete request to delete a specific user by id",
        "operationId": "deleteSpecificUser",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "user object that need get user by id",
            "required": true
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "type": "string",
            "description": "Admin Token"
          }
        ],
        "responses": {
          "200": {
            "description": "User was deleted successfully."
          },
          "401": {
            "description": "You are not authorized to use this resource"
          },
          "404": {
            "description": "Cannot delete User with this id  who is not found!"
          },
          "500": {
            "description": "internal server error"
          }
        }
      }
    },
...
  },
  "definitions": {
    "UserInput": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string",
          "required": true
        },
        "password": {
          "type": "string",
          "required": true
        }
      }
    },
...
  }
}